Move browser-related persisters in separate folder.
